fixes to UI slider and fixes to state changing which should fix most recent crash

+++ SEDaily-IOS/PodcastModel.swift
@@ -126,7 +126,7 @@ extension PodcastModel {
         }
     }
     
-    func update(currentTime: Double) {
+    func update(currentTime: Float) {
         let realm = try! Realm()
         try! realm.write {
             self.currentTime = String(currentTime)
